requirements.txt
Django==2.2.12
django-cors-headers==3.3.0
django-filter==2.2.0
django-jsonfield==1.4.0
django-rest-knox==4.1.0
django-rest-passwordreset==1.1.0
django-timezone-field==4.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
names==0.3.0
netifaces==0.10.6
pytz==2020.1
requests==2.24.0
service-identity==18.1.0
simplejwt==2.0.1
urllib3==1.25.9